umd_dining_api/requirements.txt
beautifulsoup4==4.14.3
Flask==3.1.2
flask-cors==6.0.2
pymongo[srv]==4.16.0
python-dotenv==1.2.1
requests==2.32.5
gunicorn==23.0.0